+++ Cargo.toml
@@ -12,8 +12,8 @@ epoll = ">=4.0.1"
 lazy_static = "1.4.0"
 libc = "0.2.66"
 log = { version = "0.4.10", features = ["std"] }
-qcow = { path = "qcow" }
 vhost_user_backend = { path = "vhost_user_backend"}
+vhost_user_block = { path = "vhost_user_block"}
 vhost_user_fs = { path = "vhost_user_fs"}
 vhost_user_net = { path = "vhost_user_net"}
 virtio-bindings = "0.1.0"
@@ -55,6 +55,7 @@ members = [
     "vmm",
     "vm-virtio",
     "vm-device",
+    "vhost_user_block",
     "vhost_user_backend",
     "vhost_user_fs",
     "vhost_user_net",
